+++ packages/vike-react/client/hooks/useNavigate.tsx
@@ -1,10 +1,6 @@
 export { useNavigate }
 import { navigate } from 'vike/client/router'
-
-type NavigateOptionsClient = Parameters<typeof navigate>[1]
-type NavigateOptionsServer = { statusCode?: 301 | 302 | undefined }
-type NavigateOptions = NavigateOptionsClient & NavigateOptionsServer
-type NavigateFn = (url: string, options?: NavigateOptions) => void
+import type { NavigateFn } from '../../shared/hooks/useNavigate.js'
 
 function useNavigate() {
   return navigate as NavigateFn
