+++ packages/vike-react/shared/hooks/useNavigate.tsx
@@ -1,26 +1,9 @@
-export { useNavigate }
-import { useStream } from 'react-streaming'
-import { redirect } from 'vike/abort'
+export { useNavigate, type NavigateFn }
+import { navigate } from 'vike/client/router'
 
+type NavigateOptionsClient = Parameters<typeof navigate>[1]
 type NavigateOptionsServer = { statusCode?: 301 | 302 | undefined }
+type NavigateOptions = NavigateOptionsClient & NavigateOptionsServer
+type NavigateFn = (url: string, options?: NavigateOptions) => void
 
-function useNavigate() {
-  let stream: ReturnType<typeof useStream> | null = null
-  try {
-    stream = useStream()
-  } catch (error) {
-    // Don't throw if streaming is disabled
-    // The redirect will be sent in the location header
-  }
-  const navigateFn = (url: string, options?: NavigateOptionsServer) => {
-    if (stream) {
-      // TODO: If the headers are already sent, make sure this is flushed before aborting on the server
-      // Otherwise the client will render the tree again, and navigate client-side at a later point
-      stream.injectToStream(`<script>window.location.href="${url}"</script>`)
-    }
-
-    throw redirect(url, options?.statusCode)
-  }
-
-  return navigateFn
-}
+declare function useNavigate(): NavigateFn
