+++ packages/vike-react/shared/tsconfig.json
@@ -3,6 +3,7 @@
   "compilerOptions": {
     "outDir": "../dist/",
     "target": "ES2020",
-    "module": "Node16"
+    "module": "ES2020",
+    "moduleResolution": "Node"
   }
 }
