+++ packages/vike-react/tsconfig.json
@@ -5,6 +5,9 @@
       "path": "./client/"
     },
     {
+      "path": "./node/"
+    },
+    {
       "path": "./shared/"
     }
   ]
